AM-927: Launch Darkly initial checkin.

+++ src/main/java/uk/gov/hmcts/reform/roleassignment/controller/endpoints/DeleteAssignmentController.java
@@ -1,11 +1,14 @@
 
 package uk.gov.hmcts.reform.roleassignment.controller.endpoints;
 
+import java.io.IOException;
+
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,12 +18,21 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.reform.roleassignment.domain.service.deleteroles.DeleteRoleAssignmentOrchestrator;
+import uk.gov.hmcts.reform.roleassignment.launchdarkly.LdFlagChecker;
+import uk.gov.hmcts.reform.roleassignment.util.Constants;
+import uk.gov.hmcts.reform.roleassignment.util.SecurityUtils;
 import uk.gov.hmcts.reform.roleassignment.v1.V1;
 
 
 @Api(value = "roles")
 @RestController
 public class DeleteAssignmentController {
+
+    @Autowired
+    private LdFlagChecker ldFlagChecker;
+    @Autowired
+    private SecurityUtils securityUtils;
+
     private DeleteRoleAssignmentOrchestrator deleteRoleAssignmentOrchestrator;
 
     public DeleteAssignmentController(DeleteRoleAssignmentOrchestrator deleteRoleAssignmentOrchestrator) {
@@ -89,7 +101,12 @@ public class DeleteAssignmentController {
         @RequestHeader(value = "assignerId", required = false)
             String assignerId,
         @ApiParam(value = "assignmentId", required = true)
-        @PathVariable String assignmentId) {
+        @PathVariable String assignmentId) throws IOException {
+
+        if (!ldFlagChecker
+            .verifyServiceAndFlag(securityUtils.getServiceName(), Constants.DELETE_BY_ASSIGNMENT_ID_FLAG)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Constants.ENDPOINT_NOT_AVAILABLE);
+        }
         return deleteRoleAssignmentOrchestrator.deleteRoleAssignmentByAssignmentId(assignmentId);
     }
 }
